Extract all permission gating composables and logic to PermissionGateScreen and PermissionGateViewModel
